# Optimize the parsing algorithm