@jridgewell 2018-05-11
> I'm going to push up a commit before merging that fixes the changes lines to match StandardJS style if you don't mind

Certainly. I trie to match the current style, but it was inconsistent with semis, etc.

Also note, I forgot to actually commit the maybe decode logic. It's pushed now. 😳

- - -

I have another optimization to speed up the substring+trim, which could actually be useful for a [few of your other](https://github.com/search?q=org%3Ajshttp+gather&type=Code) repos. I didn't know how to best contribute this one.