@dougwilson 2018-05-11
Oops, sorry, didn't see you pushed up more code. I'll incorporate that. I also added some benchmarks `npm run bench` to master, as I like to keep benchmarks here to help with improvements. I also ran the benchmarks you linked to and it seems the code (I was running without the decode conditional, as that's what I pulled down before you pushed it up) seems to always be worse on Node.js versions prior to 10. Can you confirm my findings? Does the split change actually regress performance in things like Node.js 6 and 8?

> I have another optimization to speed up the substring+trim, which could actually be useful for a few of your other repos. I didn't know how to best contribute this one.

I'm not sure I understand here. I'm confused by what the link is showing me, and wouldn't opening a pull request be the best method?