@jridgewell 2018-05-11
Ah man, I didn't check previous nodes. This is due to the `String.prototype.trim` calls, which aren't optimized very much.

Node 6

|  | master | indexOf | indexOf + better trim | all |
| --- | --- | --- | --- | --- |
| simple | 1,006,285 | 1,242,522 |  1,491,170 | 1,519,272 |
| decode | 883,882 | 941,025 | 1,236,415 | 1,098,289 |
| unqoute | 900,368 | 910,147 | 1,138,779 | 1,445,401 |
| 10 cookies | 127,484 | 116,429 | 138,414 | 157,773 |
| 100 cookies | 13,570 | 13,278 | 12,130 | 15,483 |


Node 8

|  | master | indexOf | indexOf + better trim | all |
| --- | --- | --- | --- | --- |
| simple | 1,499,170 | 1,666,035 | 1,535,936  | 3,420,799 |
| decode | 1,182,573 | 1,232,686 | 1,276,003 | 1,247,102 |
| unqoute | 1,283,143 | 1,453,046 | 1,470,901 | 3,105,529 |
| 10 cookies | 161,937 | 147,535 | 168,235 | 283,175 |
| 100 cookies | 14,223 | 11,949 | 15,829 | 22,697 |


Node 10

|  | master | indexOf | indexOf + better trim | all |
| --- | --- | --- | --- | --- |
| simple | 1,616,947 | 1,781,281 | 1,880,838 | 5,041,082 |
| decode | 1,168,478 | 1,315,712 | 1,323,114 | 1,196,121 |
| unqoute | 1,314,522 | 1,381,824 | 1,348,206 | 4,465,016 |
| 10 cookies | 152,325 | 171,879 | 171,892 | 387,023 |
| 100 cookies | 13,206 | 15,700 | 15,530 | 31,561 |

> I'm not sure I understand here. I'm confused by what the link is showing me, and wouldn't opening a pull request be the best method?

I can certainly include it in this PR. The link was meant that it's generally useful. You have 3 other repos that have a "parse, trim, substring" algorithm that this would solve a little better.

Personally, I would put it into a new `wsp-trim` module that this and the others would depend on and import. That way we don't duplicate the parse method 3 times.