@dougwilson 2018-05-11
The link is a search for the word "gather" and turns up results that don't even have trim in them. That's what I'm confused about.