@dougwilson 2018-05-11
So benchmarking is showing me that trim without the indexOf changes is faster than the indexOf and trim changes. Should we just land only the trim changes and not indexOf? The indexOf seems to be slowing down the gains on the majority of the versions.