@jridgewell 2018-05-11
> The link is a search for the word "gather" and turns up results that don't even have trim in them.

Oh, they're all just some variation of [parse](https://github.com/jshttp/vary/blob/71999e4e951b1edda345dc86641d0487d4e73305/index.js#L96-L123) algorithms, where they find a series of words, trim around it, and splice that out. Written with our trim:

```js
function parse (str) {
  var list = []
  
  var start = 0
  var end = str.indexOf(',');

  // gather tokens
  while (end !== -1) {
    list.push(trim(str, start, end));
    start = end + 1;
    end = str.indexOf(',', start);
  }

  // final token
  list.push(trim(str, start, str.length))

  return list
}
```

They may or may not be faster than the for-switch, but it's all the same generic algorithm.

(Also, that parse algorithm is almost byte-for-byte repeated in the three repos, which could be extracted into its own module).