@jridgewell 2018-05-11
> Yea, that's pretty much because I have no desire to maintain yet another module.

I was hesitant to just make a module because of your supported node versions. 😬

> So benchmarking is showing me that trim without the indexOf changes is faster than the indexOf and trim changes. Should we just land only the trim changes and not indexOf? The indexOf seems to be slowing down the gains on the majority of the versions.

I wouldn't. The benchmarks we've are all very short walk length to the next segment. At such short lengths, `indexOf` is pretty meh and doesn't shine.

However, for medium and longer strings (which, I intuit, is common for cookies), it's no contest. `indexOf` outshines a for-loop any day of the week.

https://jsbench.github.io/#9b5a4545655e4a1b4a4f80005668b866

Take for instance, the cookies assigned when I visit `medium.com` in incognito mode:

```
_ga=GA1.2.1666276701.1526081034; _gid=GA1.2.1030046584.1526081034; _gat=1; lightstep_guid/medium-web=370860e1f6029eff; lightstep_session_id=db23b03d7fbf9ab0; pr=2; tz=240; sz=506; __cfduid=dfcfb79745c80981519c6fcfd5f5c87c31526081032; sid=1:2hyFLAgqPBRdeOI4l9rAovdTU8mAWD7h0SrN8UGDeQd5lZ7Tp5oIUjy22Dj/nZCp; uid=lo_3ckPQk9qB79o
```

I think if we test against this string, we'll see that `indexOf` is the way to go.