@dougwilson 2018-05-11
> I think if we test against this string, we'll see that indexOf is the way to go.

Ah maybe that is it. I'll be back at a computer tomorrow or so and will check this out :+1: and update bench to be more real haha