@dougwilson 2018-05-12
So I looked at the jsbench link there and ran it on Chrome (not exactly equal to the relevant Node.js versions, but not sure yet how to run it in Node.js just yet) and tje forWalk ones were all about 2x faster than the indexOfWalk ones for the same length string. I feel like you were saying the opposite of what I'm seeing when I run this bench? Not sure what would be behind the discrepancy, but really running in web browser is not relevant here as this module targets Node.js.