@dougwilson 2018-05-12
Ok, that's cool, but why not spend the time to help make the requested changes to land instead of posting benchmarks? (1) move the trim into a module and (2) update our included benchmarks so when running we see perf improvement instead of degradation :+1: I didn't say I don't believe you, only I ran it in Chrome (just did it again) and for walk is 2x faster and it still doesn't matter since this module doesn't target my browser :)