@dougwilson 2018-05-12
And there is also the question above on if one of the changes is semver major change or no when you have a chance to look.