@jridgewell 2018-05-14
> 1. move the trim into a module
> if one of the changes is semver major change or no when you have a chance to look.

Done. Updated so it behaves exactly the same as trim.

> 2. update our included benchmarks so when running we see perf improvement instead of degradation

https://github.com/jshttp/cookie/pull/79

Full analysis at https://docs.google.com/spreadsheets/d/1rtusc7-sZ6ccIGpKKQKz70fGjrkZgQOHu2R5Go9CqY8. Reviewing each of the three changes (`indexOf`, `trim`, and `maybeDecode`), they all seem to be consistently positive changes across node versions and tests.